+++ .lintstagedrc.mjs
@@ -1,0 +1,3 @@
+export default {
+	"*.{css,scss,js,jsx,ts,tsx,json,md}": "biome format --write",
+};
